criação do protocolo e mais

+++ database/migrations/2023_06_14_173321_create_chamados_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->integer('codcli');
             $table->enum('status', ['abertura', 'chamado', 'analise de produto', 'teste de aplicacao', 'laudo', 'tratativa', 'aprovacao', 'finalizado', 'cancelado', 'reprovado']);
-            $table->integer('protocolo')->unique();
+            $table->BigInteger('protocolo')->unique();
             $table->foreign('protocolo')->references('protocolo')->on('formularios');
             $table->timestamps();
         });
