+++ database/migrations/2023_06_14_173325_create_produtos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('produtos', function (Blueprint $table) {
             $table->id();
-            $table->integer('protocolo');
+            $table->BigInteger('protocolo');
             $table->foreign('protocolo')->references('protocolo')->on('chamados');
             $table->enum('tipo', ['produto', 'tratativa']);
             $table->string('lote');
