+++ database/migrations/2023_06_14_173328_create_testes_chamados_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('testes_chamados', function (Blueprint $table) {
             $table->id();
-            $table->integer('protocolo');
+            $table->BigInteger('protocolo');
             $table->foreign('protocolo')->references('protocolo')->on('chamados');
             $table->unsignedBigInteger('id_teste');
             $table->foreign('id_teste')->references('id')->on('testes');
