+++ database/migrations/2023_06_14_173332_create_anexos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('anexos', function (Blueprint $table) {
             $table->id();
-            $table->integer('protocolo')->unique();
+            $table->BigInteger('protocolo')->unique();
             $table->foreign('protocolo')->references('protocolo')->on('chamados');
             $table->string('nome');
             $table->string('caminho');
