+++ database/migrations/2023_06_14_173341_create_log_chamados_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('log_chamados', function (Blueprint $table) {
             $table->id();
-            $table->integer('protocolo');
+            $table->BigInteger('protocolo');
             $table->foreign('protocolo')->references('protocolo')->on('chamados');
             $table->unsignedBigInteger('id_user');
             $table->foreign('id_user')->references('id')->on('users');
